+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 
+## 1.0.2
+
+- break change: rename `FloatingController.custom` to `FloatingController.transition` to better reflect its purpose and usage. 
+- break change: merge `FloatingController.route` and `FloatingController.overlay` into a single factory constructor 
+
 ## 1.0.1
 
 - support optional transition builder for `FloatingController.overlay` and `RawOverlayConfig` to enable custom transition animations for overlays.